Fix comparison and push-double instructions

+++ Altar.NET/Decomp/Decompiler.cs
@@ -540,7 +540,7 @@ namespace Altar.Decomp
                                         v = *(long*)rest;
                                         break;
                                     case DataType.String:
-                                        v = SectionReader.GetStringInfo(content, ps.ValueRest);
+                                        v = SectionReader.GetStringInfo(content, (uint)ps.ValueRest);
                                         break;
                                 }
                                 #endregion
